[Bronze II] Title: 숫자, Time: 16 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -7,19 +7,19 @@ int main()
 	ios::sync_with_stdio(0);
 	cin.tie(0);
 
-	int a, b;
+	long long a, b;
 	cin >> a >> b;
 
 
-	int Min = min(a, b);
-	int Max = max(a, b);
+	long long Min = min(a, b);
+	long long Max = max(a, b);
 
 	if (Min == Max)
 		cout << 0 << '\n';
 	else
 		cout << (Max - Min) - 1 << '\n';
 
-	for (size_t i = Min + 1; i < Max; i++)
+	for (long long i = Min + 1; i < Max; i++)
 	{
 		cout << i << " ";
 	}
